finish ECS player migration and fix map change bugs
Preserves every player component across map transitions via a new
MapWorld.ClonePlayer helper, fixing lost state (Invisible, Morph,
combat stats, SP, flags, timing) that was silently reset whenever an
entity was destroyed and recreated in the new world.

- MapChangeService: snapshots all 17 player components from the old
  world, creates the new entity via ClonePlayer, then destroys the old
  entity; also re-registers the character in its Group so the boxed
  bundle reference points at the new entity (was the NRE you saw).
- ChangeMapAsync: use mapInstance.MapInstanceId directly instead of
  routing through PositionComponent, which made teleports target the
  current map.
- Initialize miniland on character select; refactor
  IMinilandService/ICharacterInitializationService to take
  PlayerComponentBundle; replace Miniland.CharacterEntity (stale boxed
  bundle) with CharacterName, use OwnerId for ownership comparison.
- MapWorld.CreatePlayer: size 10 -> 100.

+++ src/NosCore.GameObject/Ecs/Components/ReputationComponent.cs
@@ -1,3 +1,3 @@
 namespace NosCore.GameObject.Ecs.Components;
 
-public record struct ReputationComponent(long Reputation, int Dignity, int Compliment);
+public record struct ReputationComponent(long Reputation, short Dignity, short Compliment);
