+++ src/NosCore.GameObject/Ecs/Components/SpComponent.cs
@@ -1,3 +1,3 @@
 namespace NosCore.GameObject.Ecs.Components;
 
-public record struct SpComponent(int SpCooldown, byte SpPoint, byte SpAdditionPoint);
+public record struct SpComponent(int SpCooldown, int SpPoint, int SpAdditionPoint);
