+++ src/NosCore.GameObject/Ecs/Components/VisualComponent.cs
@@ -1,3 +1,3 @@
 namespace NosCore.GameObject.Ecs.Components;
 
-public record struct VisualComponent(short Morph, byte MorphUpgrade, byte MorphDesign, byte MorphBonus, bool NoAttack, bool NoMove, bool IsSitting);
+public record struct VisualComponent(short Morph, byte MorphUpgrade, short MorphDesign, byte MorphBonus, bool NoAttack, bool NoMove, bool IsSitting);
